Updated Naive Bayes

+++ src/utils/Utility.java
@@ -62,7 +62,7 @@ public class Utility {
 		}
 		
 		int numUsers = recommendations.keySet().size();
-		return (overallAccuracy * 100) / numUsers;
+		return (overallAccuracy * 100) / (double) numUsers;
 	}
 	
 	public static Algorithm getAlgorithmInstance(String algoName, int numSongsToRecommend)
@@ -194,6 +194,7 @@ public class Utility {
 				{
 					List<String> listenersList = songMap.get(songID).getListenersList();
 					listenersList.add(userID);
+					//System.out.println(songMap.get(songID).getListenersList().size());
 				}
 				else
 				{
